Refactoring client
