+++ README.md
@@ -10,7 +10,7 @@ Currently, this library supports database operations only
 // Starting the client
 SupabaseRestClient supabase = new SupabaseRestClient(YOUR_SUPABASE_URL, YOUR_SUPABASE_KEY);
 // Inserting a new user to the table users
-supabase.insert("users", Insert.row().column("email", "user@email.com")
+supabase.databse().insert("users", Insert.row().column("email", "user@email.com")
                                      .column("username", "user123"));
 ```
