+++ src/main/java/com/harium/supabase/RequestDecorator.java
@@ -1,0 +1,9 @@
+package com.harium.supabase;
+
+import okhttp3.Request;
+
+public interface RequestDecorator {
+
+    Request.Builder decorate(Request.Builder builder);
+
+}
