+++ src/main/java/com/harium/supabase/SupabaseClient.java
@@ -1,58 +1,28 @@
 package com.harium.supabase;
 
-import com.harium.postgrest.Condition;
-import com.harium.postgrest.PostgrestClient;
-import com.sun.corba.se.impl.util.Version;
-import okhttp3.HttpUrl;
+import com.harium.supabase.database.DatabaseClient;
 import okhttp3.Request;
 
-public class SupabaseClient extends PostgrestClient {
+public class SupabaseClient implements RequestDecorator {
 
     private final String supabaseKey;
 
+    private final DatabaseClient databaseClient;
+
     public SupabaseClient(String supabaseUrl, String supabaseKey) {
-        super(supabaseUrl);
         this.supabaseKey = supabaseKey;
+        databaseClient = new DatabaseClient(supabaseUrl, this);
     }
 
-    protected HttpUrl buildTableUrl(String table) {
-        return new HttpUrl.Builder()
-                .scheme("https")
-                .host(baseUrl)
-                .addPathSegment("rest")
-                .addPathSegment("v1")
-                .addPathSegment(table)
-                .build();
+    public DatabaseClient database() {
+        return databaseClient;
     }
 
-    protected HttpUrl buildTableUrl(String table, Condition condition) {
-        return new HttpUrl.Builder()
-                .scheme("https")
-                .host(baseUrl)
-                .addPathSegment("rest")
-                .addPathSegment("v1")
-                .addPathSegment(table)
-                .addQueryParameter(condition.getQueryParam(), condition.getQueryValue())
-                .build();
-    }
-
-    protected Request.Builder buildRequest(String table) {
-        HttpUrl httpUrl = buildTableUrl(table);
-
-        return new Request.Builder()
+    @Override
+    public Request.Builder decorate(Request.Builder builder) {
+        return builder
                 .addHeader("apikey", supabaseKey)
                 .addHeader("Authorization", "Bearer " + supabaseKey)
-                .addHeader("X-Client-Info", "supabase-java/" + Version.VERSION)
-                .url(httpUrl);
-    }
-
-    protected Request.Builder buildRequest(String table, Condition condition) {
-        HttpUrl httpUrl = buildTableUrl(table, condition);
-
-        return new Request.Builder()
-                .addHeader("apikey", supabaseKey)
-                .addHeader("Authorization", "Bearer " + supabaseKey)
-                .addHeader("X-Client-Info", "supabase-java/" + Version.VERSION)
-                .url(httpUrl);
+                .addHeader("X-Client-Info", "supabase-java/" + Version.VERSION);
     }
 }
