+++ src/main/java/com/harium/supabase/Version.java
@@ -1,0 +1,7 @@
+package com.harium.supabase;
+
+public class Version {
+
+    public static final String VERSION = "0.0.1";
+
+}
